Get events

Get event ids given a free-text query

Usage

get_events(
  query = NULL,
  ...,
  sort_by = "startDate",
  ascending = FALSE,
  has_series = TRUE,
  n_results = 50
)

Arguments

query

Text description of event, e.g. "VALORANT Champions 2022 Playoffs"

...

Currently un-used

sort_by, ascending

How to return the results

has_series

Whether or not to return events with completed series

n_results

Number of results to return. API tries to respect n_results even if event name doesn't match query. 50 is default of the API.

Value

a data.frame

Details

Note that sometimes the query doesn't seem to work. Results will be returned, but they are a default set of results.

Examples

# \donttest{
events <- get_events("VALORANT Champions 2022")
dplyr::glimpse(events)
#> Rows: 50
#> Columns: 30
#> $ id                <int> 3306, 3169, 3168, 3167, 2927, 3018, 2909, 2908, 2899…
#> $ name              <chr> "Valorant Champions Tour OFF//SEASON 2023", "VALORAN…
#> $ shortName         <chr> "VCT OFF//SEASON 2023", "Playoffs", "Group Stage", "…
#> $ description       <chr> "Part of the VCT OFF//SEASON, Riot's 2023 post-seaso…
#> $ logoUrl           <chr> NA, NA, NA, "https://i.imgur.com/bIMOlw9.webp", NA, …
#> $ regionId          <int> 7, 7, 7, 7, 3, 2, 3, 3, 3, 2, 1, 3, 2, 2, 2, 2, 1, 1…
#> $ countryId         <int> 252, 226, 226, 226, 44, 226, 43, 44, 113, 226, 80, 1…
#> $ startDate         <chr> "2023-09-07T15:25:06.000Z", "2023-08-16T04:49:31.000…
#> $ endDate           <chr> "2023-11-05T15:25:06.000Z", "2023-08-27T04:49:31.000…
#> $ prizePool         <int> NA, NA, NA, NA, NA, 3500, NA, NA, NA, NA, NA, NA, 10…
#> $ prizePoolCurrency <chr> "USD", "USD", "USD", "USD", "USD", "USD", "USD", "US…
#> $ url               <chr> "https://vct.gg", "https://liquipedia.net/valorant/V…
#> $ imageUrl          <chr> NA, NA, NA, "https://i.imgur.com/bIMOlw9.webp", NA, …
#> $ winnerStageCount  <int> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, …
#> $ loserStageCount   <int> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, …
#> $ live              <lgl> FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FAL…
#> $ rank              <int> 5, 9, 9, 9, 7, 3, 5, 7, 1, 9, 9, 9, 5, 3, 9, 8, 8, 9…
#> $ pmtJson           <lgl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, …
#> $ parent            <lgl> TRUE, FALSE, FALSE, TRUE, FALSE, FALSE, FALSE, TRUE,…
#> $ parentId          <int> NA, 3167, 3167, 2190, 2908, NA, 2908, 2918, 2535, 25…
#> $ childLabel        <chr> NA, "Playoffs", "Group Stage", "VALORANT Champions 2…
#> $ keywords          <list> [], [], [], [], [], [], [], [], [], [], [], [], [],…
#> $ slug              <chr> "valorant-champions-tour-off-season-2023", "valorant…
#> $ seriesCount       <int> 0, 14, 20, 0, 14, 3, 33, 0, 1, 11, 10, 10, 38, 98, 0…
#> $ importance        <int> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0…
#> $ type              <chr> "circuit", "bracket", "bracket", "event", "bracket",…
#> $ vctRegions        <list> <"EMEA", "AMERICAS", "PACIFIC", "CHINA">, <"AMERICA…
#> $ divisions         <list> <"T3", "GC">, "VCT", "VCT", "VCT", <"VCT", "VCL", N…
#> $ region            <df[,4]> <data.frame[26 x 4]>
#> $ country           <df[,6]> <data.frame[26 x 6]>
# }