Get matches

Get match ids for a given series

Usage

get_matches(series_id)

Arguments

series_id

Series ID

Value

a list with the following elements: id, event_id, team1id, team2id, start_date, best_of, stage, bracket, completed, live, win_condition, vlr_id, vod_url, ggbet_id, pmt_status, pmt_reddit_url, pmt_json, event_name, event_slug, event_logo_url, parent_event_id, parent_event_name, parent_event_slug, team1, team2, matches, stats, player_stats.

Examples

# \donttest{
matches <- get_matches(35225) ## Optic vs. Liquid
matches$matches
#>      id patchId seriesId seriesMatchNumber                               riotId
#> 1 79016      26    35225                 1 a90c6bfb-e656-4d0f-ab32-fa1c275374b4
#> 2 79017      26    35225                 2 f2c5ed14-6718-4219-9898-2fd8802f587f
#> 3 79018      26    35225                 3 15ef3a70-7e55-4034-9cae-f6a349942daf
#>   mapId                startDate lengthMillis completed
#> 1     8 2022-09-10T13:57:37.704Z      3708997      TRUE
#> 2     3 2022-09-10T15:08:28.124Z      3338271      TRUE
#> 3     1 2022-09-10T16:15:29.650Z      3238467      TRUE
#>                           riotSeasonId
#> 1 3e47230a-463c-a301-eb7d-67bb60357d4f
#> 2 3e47230a-463c-a301-eb7d-67bb60357d4f
#> 3 3e47230a-463c-a301-eb7d-67bb60357d4f
#>                                       riotGameMode  region  live readyToDisplay
#> 1 /Game/GameModes/Bomb/BombGameMode.BombGameMode_C esports FALSE           TRUE
#> 2 /Game/GameModes/Bomb/BombGameMode.BombGameMode_C esports FALSE           TRUE
#> 3 /Game/GameModes/Bomb/BombGameMode.BombGameMode_C esports FALSE           TRUE
#>   attackingFirstTeamNumber redTeamNumber winningTeamNumber winCondition
#> 1                        1             1                 2     standard
#> 2                        1             1                 1     standard
#> 3                        1             1                 1     standard
#>   teamsInvertedInStream team1Score team2Score vlrId
#> 1                 FALSE          7         13 95116
#> 2                 FALSE         13          8 95117
#> 3                 FALSE         13          8 95118
#>                                                vodUrl map.id map.name
#> 1 https://www.twitch.tv/videos/1587106175?t=01h16m05s      8   Breeze
#> 2 https://www.twitch.tv/videos/1587106175?t=02h13m04s      3     Bind
#> 3 https://www.twitch.tv/videos/1587106175?t=03h18m54s      1   Ascent
#>                   map.riotId
#> 1 /Game/Maps/Foxtrot/Foxtrot
#> 2 /Game/Maps/Duality/Duality
#> 3   /Game/Maps/Ascent/Ascent
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                   players
#> 1     79016, 79016, 79016, 79016, 79016, 79016, 79016, 79016, 79016, 79016, 2716, 2841, 398, 10211, 2483, 3211, 1690, 17788, 750, 2056, 1, 1, 1, 2, 1, 1, 2, 2, 2, 2, 17, 6, 1, 17, 4, 18, 7, 6, 4, 16, 2716, 2841, 398, 10211, 2483, 3211, 1690, 17788, 750, 2056, yay, FNS, Marved, Jamppi, crashies, Victor, ScreaM, Nivera, dimasick, soulcas, Jaccob , Pujan , Jimmy , Elias, Austin, Victor, Adil, Nabil, Dmitriy , Dom, Whiteaker, Mehta, Nguyen, Olkkonen, Roberts, Wong, Benrlitom, Benrlitom, Matvienko, Sulcas, Jaccob "yay" Whiteaker is an American player who is currently playing for Andbox., Pujan "FNS" Mehta (born March 19, 1992) is a Canadian player who currently plays for Team Envy. He is a former professional Counter-Strike: Global Offensive player., Jimmy "Marved" Nguyen (born February 24, 2000) is a Canadian player who currently plays for FaZe Clan. He is a retired professional Counter-Strike: Global Offensive player., NA, Austin "crashies" Roberts (born December 20, 1997) is an American player who currently plays for Team Envy., Victor "food" Wong is an American player who currently plays for Team Envy., Adil "ScreaM" Benrlitom (born July 2, 1994) is a Belgian player who currently plays for Team Liquid. He is an ex-professional Counter-Strike: Global Offensive player who played for various notable teams., NA, NA, Dom "soulcas" Sulcas (born January 2, 2000) is a British player who currently plays for Team Liquid., 226, 38, 38, 72, 226, 226, 21, 21, 109, 225, NA, https://www.instagram.com/pujmehta/, NA, NA, NA, https://www.instagram.com/t1_food/, https://www.instagram.com/screamdak1ng/, NA, NA, https://www.instagram.com/dommsulcas/, NA, https://liquipedia.net/valorant/FNS, https://liquipedia.net/valorant/Marved, NA, https://liquipedia.net/valorant/Crashies, https://liquipedia.net/valorant/Food, https://liquipedia.net/valorant/ScreaM, NA, NA, https://liquipedia.net/valorant/Soulcas, https://www.twitch.tv/yayisreal, https://www.twitch.tv/goFNS, https://www.twitch.tv/marved6, https://www.twitch.tv/superjamppi, https://twitter.com/Crashies, https://www.twitch.tv/t1food, https://twitter.com/ScreaM_, https://www.twitch.tv/nivera, https://www.twitch.tv/dimasickcsgo, https://www.twitch.tv/soulcas_, https://twitter.com/yayFPS, https://twitter.com/fns, https://twitter.com/Marved6, https://twitter.com/Jamppicsgo, https://www.twitch.tv/crashies, https://twitter.com/victorwong, https://www.twitch.tv/ScreaM, https://twitter.com/Nivera__, https://twitter.com/dimasickCSGO, https://twitter.com/soulcas_, NA, https://www.youtube.com/user/GOfns, NA, NA, https://www.youtube.com/user/MrCrashies, NA, https://www.youtube.com/ScreaMy, NA, NA, https://www.youtube.com/soulcas, https://i.imgur.com/831bo4j.webp, https://i.imgur.com/2M0G4Qg.webp, https://i.imgur.com/DYwGgCi.png, https://i.imgur.com/KXlIwbu.png, https://i.imgur.com/88KuUGZ.png, https://i.imgur.com/8i6NjK7.png, https://i.imgur.com/t7VEMDo.png, https://i.imgur.com/ly2IzwA.png, https://i.imgur.com/8P85myq.png, https://i.imgur.com/kDZQNoq.png, q90rD7XX8VW2aeoaLn9l, sHqZJPoNvD0atH19gnki, 7RC8SKxVDAcCRVyGyT6J, NA, lg8vswOSMGDhGvsYQJk6, zvj03xtcCY3tnh7zDnHR, WwCfei085XGlltIHO1T2, NA, EKfbL3XJAB7OylWeBWIq, da3qjXe98ssxE3g0FtPa, 4qhP3SAGywYPl3UYacEuagImQvGcL9dJLvk34dpxSm4MfHHLmOR8yOY-KkJKKPONEtj9KUMSThIRCA, c9eNTnnz4i_dsql4LmZ39TenkdTlpgMmW-n2FVWj7vnBA8Pcob84FfqOcPwiqGUaBjVjgIHdRGHlAw, pPlpY4bXIyRj-arVkILGP4rSc92UNgSrPOq0L0ZDbgaoHpGXKTTJnS_-zeL2pmcx3swavPS8uDUnLA, EtU32hXrPx92J0MvRAbt8kxyJdHn9aL9LJ4s1dtFXYWGha_ZljRPNHIx2lDj7ijAjhvTN7-x2ed_UQ, GcPwuG7vMwo9wHwd7VvAj4PNDxJE_Aa9REuzBgsR-8_xUXFXElZpSF5hFeVZl4IIcFhR-B4OI2ZpYw, LBswzTGyy7DjId6aY6bKdeiO2qMukE3eAaagTpxnST8doiv6xQA0BGWqbx5CSeLTCqXBzAGnStfNoA, iF_207UuSo20ScqagkRBjnAgmQwFYt-z4pxKvoze8vwZdT1RxR3qLuEvWBxZy8qW0GuKFNbu2Aa9PQ, -Wnr2eYq3J5U2UviHSnhU0yP9zDVbAV7YAZVOQi2USklvyaJYIhWfB6wtZf_jHl8GNx_VAfwmrg1eQ, RTQNP6qmFCbcbbZNssKDqAUfd0R94H1aud3BxaNh-N2qbk-j9ALdo3Du7J2znwvEEQ20h9XijBAZVQ, NBw5-ddKRfga3yMIJ31cqv1yXQsqVXEy_JbwFqnq3nxypexaF4rFFWXfdtf8QsAojlhSpYEF9AMSqw, A_73nV2Y_W1Ey9ob7_eI59p5k2hQNK1WiP-pD5CauiXnCyFq1lyWa-xX8YY42B2doKAkdWahvilqPA, j9oNuqlN0lL17LenCF5fnPsbSOGuzBeLIJksV-wvAFHoJaOsjPKwrBsKfjQC7mJELX6rEWSToNNjkw, m6DXq5RW43vAmdioDFMB3FyJ1N9vbI8Y1wMoBTsE-1MVR-bnogBaABfMWY9KkkGytjBRnIAK-GzpMQ, q3Ygpe4ufLYQ9qBIzWrPb-fkaHquWJ_Isz82VkQsSJ-hPI46mHBvn4QPLTrdtcHCMZUC7QhGc7XAYw, Tn7RPIRm2aZbnCSe8tKPgr2K9iYQSqkgG5-6gdgxylfZaPLszTuC8Lv-k2fiFP1Ro-Dw0vSMyHThvQ, PmdJJFaKylfhH5_Heu7XtbmRT7WgGhfYwjKIt6KeUjli8up9vOuLxHIF6-eHuax_yP_AzSWDuIOhtA, j0f-p41ef65Kw61Ulc5xh08GB5bNX4DIHH8yul18_b0_TkLUOo7Tr6nWkBOtm4yn-JA3ns8-5yolIA, zoopRMotdtX3SD638hT4e96onSEQZfPxHPhuo-isNXvYUjyylfwieqc3LfFKdZeyNCdKGQOpfy4KVQ, KdKTppy2q8B0J9_9KJFt5qqaycy_RCrPjOBzE1KoTIRoHVWUjRl4UpjiXPaH9X3JX17sLn8_ucpPEA, M4IVHDXbfPP-2IVCqV4v5s_WUcolEP8W9jD945TfgVvMmUFTvGG9HYe3RxzQqGVQm9B2Qe2D5ZCElg, _ENN9JICMNepnUGxOgyYwutjke6AFe897QtJfpf-vXByLXD9a8tOEJYAKB_Zmjd8KMQv0TBN1XQwEw
#> 2      79017, 79017, 79017, 79017, 79017, 79017, 79017, 79017, 79017, 79017, 1690, 2056, 398, 17788, 10211, 2483, 750, 3211, 2841, 2716, 2, 2, 1, 2, 2, 1, 2, 1, 1, 1, 2, 13, 8, 6, 9, 13, 8, 2, 6, 17, 1690, 2056, 398, 17788, 10211, 2483, 750, 3211, 2841, 2716, ScreaM, soulcas, Marved, Nivera, Jamppi, crashies, dimasick, Victor, FNS, yay, Adil, Dom, Jimmy , Nabil, Elias, Austin, Dmitriy , Victor, Pujan , Jaccob , Benrlitom, Sulcas, Nguyen, Benrlitom, Olkkonen, Roberts, Matvienko, Wong, Mehta, Whiteaker, Adil "ScreaM" Benrlitom (born July 2, 1994) is a Belgian player who currently plays for Team Liquid. He is an ex-professional Counter-Strike: Global Offensive player who played for various notable teams., Dom "soulcas" Sulcas (born January 2, 2000) is a British player who currently plays for Team Liquid., Jimmy "Marved" Nguyen (born February 24, 2000) is a Canadian player who currently plays for FaZe Clan. He is a retired professional Counter-Strike: Global Offensive player., NA, NA, Austin "crashies" Roberts (born December 20, 1997) is an American player who currently plays for Team Envy., NA, Victor "food" Wong is an American player who currently plays for Team Envy., Pujan "FNS" Mehta (born March 19, 1992) is a Canadian player who currently plays for Team Envy. He is a former professional Counter-Strike: Global Offensive player., Jaccob "yay" Whiteaker is an American player who is currently playing for Andbox., 21, 225, 38, 21, 72, 226, 109, 226, 38, 226, https://www.instagram.com/screamdak1ng/, https://www.instagram.com/dommsulcas/, NA, NA, NA, NA, NA, https://www.instagram.com/t1_food/, https://www.instagram.com/pujmehta/, NA, https://liquipedia.net/valorant/ScreaM, https://liquipedia.net/valorant/Soulcas, https://liquipedia.net/valorant/Marved, NA, NA, https://liquipedia.net/valorant/Crashies, NA, https://liquipedia.net/valorant/Food, https://liquipedia.net/valorant/FNS, NA, https://twitter.com/ScreaM_, https://www.twitch.tv/soulcas_, https://www.twitch.tv/marved6, https://www.twitch.tv/nivera, https://www.twitch.tv/superjamppi, https://twitter.com/Crashies, https://www.twitch.tv/dimasickcsgo, https://www.twitch.tv/t1food, https://www.twitch.tv/goFNS, https://www.twitch.tv/yayisreal, https://www.twitch.tv/ScreaM, https://twitter.com/soulcas_, https://twitter.com/Marved6, https://twitter.com/Nivera__, https://twitter.com/Jamppicsgo, https://www.twitch.tv/crashies, https://twitter.com/dimasickCSGO, https://twitter.com/victorwong, https://twitter.com/fns, https://twitter.com/yayFPS, https://www.youtube.com/ScreaMy, https://www.youtube.com/soulcas, NA, NA, NA, https://www.youtube.com/user/MrCrashies, NA, NA, https://www.youtube.com/user/GOfns, NA, https://i.imgur.com/t7VEMDo.png, https://i.imgur.com/kDZQNoq.png, https://i.imgur.com/DYwGgCi.png, https://i.imgur.com/ly2IzwA.png, https://i.imgur.com/KXlIwbu.png, https://i.imgur.com/88KuUGZ.png, https://i.imgur.com/8P85myq.png, https://i.imgur.com/8i6NjK7.png, https://i.imgur.com/2M0G4Qg.webp, https://i.imgur.com/831bo4j.webp, WwCfei085XGlltIHO1T2, da3qjXe98ssxE3g0FtPa, 7RC8SKxVDAcCRVyGyT6J, NA, NA, lg8vswOSMGDhGvsYQJk6, EKfbL3XJAB7OylWeBWIq, zvj03xtcCY3tnh7zDnHR, sHqZJPoNvD0atH19gnki, q90rD7XX8VW2aeoaLn9l, m6DXq5RW43vAmdioDFMB3FyJ1N9vbI8Y1wMoBTsE-1MVR-bnogBaABfMWY9KkkGytjBRnIAK-GzpMQ, q3Ygpe4ufLYQ9qBIzWrPb-fkaHquWJ_Isz82VkQsSJ-hPI46mHBvn4QPLTrdtcHCMZUC7QhGc7XAYw, M4IVHDXbfPP-2IVCqV4v5s_WUcolEP8W9jD945TfgVvMmUFTvGG9HYe3RxzQqGVQm9B2Qe2D5ZCElg, _ENN9JICMNepnUGxOgyYwutjke6AFe897QtJfpf-vXByLXD9a8tOEJYAKB_Zmjd8KMQv0TBN1XQwEw, GcPwuG7vMwo9wHwd7VvAj4PNDxJE_Aa9REuzBgsR-8_xUXFXElZpSF5hFeVZl4IIcFhR-B4OI2ZpYw, LBswzTGyy7DjId6aY6bKdeiO2qMukE3eAaagTpxnST8doiv6xQA0BGWqbx5CSeLTCqXBzAGnStfNoA, Tn7RPIRm2aZbnCSe8tKPgr2K9iYQSqkgG5-6gdgxylfZaPLszTuC8Lv-k2fiFP1Ro-Dw0vSMyHThvQ, PmdJJFaKylfhH5_Heu7XtbmRT7WgGhfYwjKIt6KeUjli8up9vOuLxHIF6-eHuax_yP_AzSWDuIOhtA, iF_207UuSo20ScqagkRBjnAgmQwFYt-z4pxKvoze8vwZdT1RxR3qLuEvWBxZy8qW0GuKFNbu2Aa9PQ, -Wnr2eYq3J5U2UviHSnhU0yP9zDVbAV7YAZVOQi2USklvyaJYIhWfB6wtZf_jHl8GNx_VAfwmrg1eQ, RTQNP6qmFCbcbbZNssKDqAUfd0R94H1aud3BxaNh-N2qbk-j9ALdo3Du7J2znwvEEQ20h9XijBAZVQ, NBw5-ddKRfga3yMIJ31cqv1yXQsqVXEy_JbwFqnq3nxypexaF4rFFWXfdtf8QsAojlhSpYEF9AMSqw, j0f-p41ef65Kw61Ulc5xh08GB5bNX4DIHH8yul18_b0_TkLUOo7Tr6nWkBOtm4yn-JA3ns8-5yolIA, zoopRMotdtX3SD638hT4e96onSEQZfPxHPhuo-isNXvYUjyylfwieqc3LfFKdZeyNCdKGQOpfy4KVQ, KdKTppy2q8B0J9_9KJFt5qqaycy_RCrPjOBzE1KoTIRoHVWUjRl4UpjiXPaH9X3JX17sLn8_ucpPEA, A_73nV2Y_W1Ey9ob7_eI59p5k2hQNK1WiP-pD5CauiXnCyFq1lyWa-xX8YY42B2doKAkdWahvilqPA, j9oNuqlN0lL17LenCF5fnPsbSOGuzBeLIJksV-wvAFHoJaOsjPKwrBsKfjQC7mJELX6rEWSToNNjkw, pPlpY4bXIyRj-arVkILGP4rSc92UNgSrPOq0L0ZDbgaoHpGXKTTJnS_-zeL2pmcx3swavPS8uDUnLA, EtU32hXrPx92J0MvRAbt8kxyJdHn9aL9LJ4s1dtFXYWGha_ZljRPNHIx2lDj7ijAjhvTN7-x2ed_UQ, 4qhP3SAGywYPl3UYacEuagImQvGcL9dJLvk34dpxSm4MfHHLmOR8yOY-KkJKKPONEtj9KUMSThIRCA, c9eNTnnz4i_dsql4LmZ39TenkdTlpgMmW-n2FVWj7vnBA8Pcob84FfqOcPwiqGUaBjVjgIHdRGHlAw
#> 3 79018, 79018, 79018, 79018, 79018, 79018, 79018, 79018, 79018, 79018, 2056, 10211, 3211, 1690, 750, 2483, 2841, 398, 2716, 17788, 2, 2, 1, 2, 2, 1, 1, 1, 1, 2, 19, 17, 16, 7, 11, 4, 19, 15, 17, 16, 2056, 10211, 3211, 1690, 750, 2483, 2841, 398, 2716, 17788, soulcas, Jamppi, Victor, ScreaM, dimasick, crashies, FNS, Marved, yay, Nivera, Dom, Elias, Victor, Adil, Dmitriy , Austin, Pujan , Jimmy , Jaccob , Nabil, Sulcas, Olkkonen, Wong, Benrlitom, Matvienko, Roberts, Mehta, Nguyen, Whiteaker, Benrlitom, Dom "soulcas" Sulcas (born January 2, 2000) is a British player who currently plays for Team Liquid., NA, Victor "food" Wong is an American player who currently plays for Team Envy., Adil "ScreaM" Benrlitom (born July 2, 1994) is a Belgian player who currently plays for Team Liquid. He is an ex-professional Counter-Strike: Global Offensive player who played for various notable teams., NA, Austin "crashies" Roberts (born December 20, 1997) is an American player who currently plays for Team Envy., Pujan "FNS" Mehta (born March 19, 1992) is a Canadian player who currently plays for Team Envy. He is a former professional Counter-Strike: Global Offensive player., Jimmy "Marved" Nguyen (born February 24, 2000) is a Canadian player who currently plays for FaZe Clan. He is a retired professional Counter-Strike: Global Offensive player., Jaccob "yay" Whiteaker is an American player who is currently playing for Andbox., NA, 225, 72, 226, 21, 109, 226, 38, 38, 226, 21, https://www.instagram.com/dommsulcas/, NA, https://www.instagram.com/t1_food/, https://www.instagram.com/screamdak1ng/, NA, NA, https://www.instagram.com/pujmehta/, NA, NA, NA, https://liquipedia.net/valorant/Soulcas, NA, https://liquipedia.net/valorant/Food, https://liquipedia.net/valorant/ScreaM, NA, https://liquipedia.net/valorant/Crashies, https://liquipedia.net/valorant/FNS, https://liquipedia.net/valorant/Marved, NA, NA, https://www.twitch.tv/soulcas_, https://www.twitch.tv/superjamppi, https://www.twitch.tv/t1food, https://twitter.com/ScreaM_, https://www.twitch.tv/dimasickcsgo, https://twitter.com/Crashies, https://www.twitch.tv/goFNS, https://www.twitch.tv/marved6, https://www.twitch.tv/yayisreal, https://www.twitch.tv/nivera, https://twitter.com/soulcas_, https://twitter.com/Jamppicsgo, https://twitter.com/victorwong, https://www.twitch.tv/ScreaM, https://twitter.com/dimasickCSGO, https://www.twitch.tv/crashies, https://twitter.com/fns, https://twitter.com/Marved6, https://twitter.com/yayFPS, https://twitter.com/Nivera__, https://www.youtube.com/soulcas, NA, NA, https://www.youtube.com/ScreaMy, NA, https://www.youtube.com/user/MrCrashies, https://www.youtube.com/user/GOfns, NA, NA, NA, https://i.imgur.com/kDZQNoq.png, https://i.imgur.com/KXlIwbu.png, https://i.imgur.com/8i6NjK7.png, https://i.imgur.com/t7VEMDo.png, https://i.imgur.com/8P85myq.png, https://i.imgur.com/88KuUGZ.png, https://i.imgur.com/2M0G4Qg.webp, https://i.imgur.com/DYwGgCi.png, https://i.imgur.com/831bo4j.webp, https://i.imgur.com/ly2IzwA.png, da3qjXe98ssxE3g0FtPa, NA, zvj03xtcCY3tnh7zDnHR, WwCfei085XGlltIHO1T2, EKfbL3XJAB7OylWeBWIq, lg8vswOSMGDhGvsYQJk6, sHqZJPoNvD0atH19gnki, 7RC8SKxVDAcCRVyGyT6J, q90rD7XX8VW2aeoaLn9l, NA, M4IVHDXbfPP-2IVCqV4v5s_WUcolEP8W9jD945TfgVvMmUFTvGG9HYe3RxzQqGVQm9B2Qe2D5ZCElg, _ENN9JICMNepnUGxOgyYwutjke6AFe897QtJfpf-vXByLXD9a8tOEJYAKB_Zmjd8KMQv0TBN1XQwEw, iF_207UuSo20ScqagkRBjnAgmQwFYt-z4pxKvoze8vwZdT1RxR3qLuEvWBxZy8qW0GuKFNbu2Aa9PQ, -Wnr2eYq3J5U2UviHSnhU0yP9zDVbAV7YAZVOQi2USklvyaJYIhWfB6wtZf_jHl8GNx_VAfwmrg1eQ, A_73nV2Y_W1Ey9ob7_eI59p5k2hQNK1WiP-pD5CauiXnCyFq1lyWa-xX8YY42B2doKAkdWahvilqPA, j9oNuqlN0lL17LenCF5fnPsbSOGuzBeLIJksV-wvAFHoJaOsjPKwrBsKfjQC7mJELX6rEWSToNNjkw, m6DXq5RW43vAmdioDFMB3FyJ1N9vbI8Y1wMoBTsE-1MVR-bnogBaABfMWY9KkkGytjBRnIAK-GzpMQ, q3Ygpe4ufLYQ9qBIzWrPb-fkaHquWJ_Isz82VkQsSJ-hPI46mHBvn4QPLTrdtcHCMZUC7QhGc7XAYw, j0f-p41ef65Kw61Ulc5xh08GB5bNX4DIHH8yul18_b0_TkLUOo7Tr6nWkBOtm4yn-JA3ns8-5yolIA, zoopRMotdtX3SD638hT4e96onSEQZfPxHPhuo-isNXvYUjyylfwieqc3LfFKdZeyNCdKGQOpfy4KVQ, KdKTppy2q8B0J9_9KJFt5qqaycy_RCrPjOBzE1KoTIRoHVWUjRl4UpjiXPaH9X3JX17sLn8_ucpPEA, RTQNP6qmFCbcbbZNssKDqAUfd0R94H1aud3BxaNh-N2qbk-j9ALdo3Du7J2znwvEEQ20h9XijBAZVQ, NBw5-ddKRfga3yMIJ31cqv1yXQsqVXEy_JbwFqnq3nxypexaF4rFFWXfdtf8QsAojlhSpYEF9AMSqw, pPlpY4bXIyRj-arVkILGP4rSc92UNgSrPOq0L0ZDbgaoHpGXKTTJnS_-zeL2pmcx3swavPS8uDUnLA, EtU32hXrPx92J0MvRAbt8kxyJdHn9aL9LJ4s1dtFXYWGha_ZljRPNHIx2lDj7ijAjhvTN7-x2ed_UQ, GcPwuG7vMwo9wHwd7VvAj4PNDxJE_Aa9REuzBgsR-8_xUXFXElZpSF5hFeVZl4IIcFhR-B4OI2ZpYw, LBswzTGyy7DjId6aY6bKdeiO2qMukE3eAaagTpxnST8doiv6xQA0BGWqbx5CSeLTCqXBzAGnStfNoA, 4qhP3SAGywYPl3UYacEuagImQvGcL9dJLvk34dpxSm4MfHHLmOR8yOY-KkJKKPONEtj9KUMSThIRCA, c9eNTnnz4i_dsql4LmZ39TenkdTlpgMmW-n2FVWj7vnBA8Pcob84FfqOcPwiqGUaBjVjgIHdRGHlAw, Tn7RPIRm2aZbnCSe8tKPgr2K9iYQSqkgG5-6gdgxylfZaPLszTuC8Lv-k2fiFP1Ro-Dw0vSMyHThvQ, PmdJJFaKylfhH5_Heu7XtbmRT7WgGhfYwjKIt6KeUjli8up9vOuLxHIF6-eHuax_yP_AzSWDuIOhtA
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                              stats
#> 1           79016, 79016, 79016, 79016, 79016, 79016, 79016, 79016, 79016, 79016, 398, 750, 1690, 2056, 2483, 2716, 2841, 3211, 10211, 17788, 2133, 3563, 5390, 5900, 3900, 5921, 4595, 2757, 4497, 2633, 20, 20, 20, 20, 20, 20, 20, 20, 20, 20, 7, 13, 19, 22, 14, 20, 16, 9, 16, 10, 17, 14, 12, 14, 14, 17, 15, 17, 12, 14, 6, 6, 5, 12, 4, 4, 8, 2, 6, 3, 3708997, 3708997, 3708997, 3708997, 3708997, 3708997, 3708997, 3708997, 3708997, 3708997, -199.794816037198164000, 4.132774377647329800, 135.866377371564048700, 3.3347819693026360500, -27.029158670031090900, 23.7114184424397036500, 8.935876919853467200, -127.252259677026920200, 160.21057181516508500, 17.884433488283904700
#> 2 79017, 79017, 79017, 79017, 79017, 79017, 79017, 79017, 79017, 79017, 398, 750, 1690, 2056, 2483, 2716, 2841, 3211, 10211, 17788, 5715, 3575, 5606, 4660, 4827, 5432, 3723, 4206, 2520, 3205, 21, 21, 21, 21, 21, 21, 21, 21, 21, 21, 20, 13, 17, 18, 20, 19, 11, 15, 8, 9, 14, 18, 16, 17, 13, 14, 12, 12, 18, 16, 11, 2, 10, 5, 7, 4, 12, 4, 4, 6, 3338271, 3338271, 3338271, 3338271, 3338271, 3338271, 3338271, 3338271, 3338271, 3338271, -11.2794047732629927200, -50.5086913697045379200, 56.654069338426370500, 112.459219166317556600, 53.238535026040709400, 110.4984650496974489200, -101.751089475358066000, 55.182994373692738600, -188.438793920011166400, -30.8534622214930918300
#> 3           79018, 79018, 79018, 79018, 79018, 79018, 79018, 79018, 79018, 79018, 398, 750, 1690, 2056, 2483, 2716, 2841, 3211, 10211, 17788, 4808, 2127, 4271, 2917, 2371, 7849, 2729, 4653, 4727, 3566, 21, 21, 21, 21, 21, 21, 21, 21, 21, 21, 16, 7, 14, 9, 7, 30, 8, 17, 15, 13, 9, 15, 16, 16, 12, 9, 14, 14, 16, 15, 11, 4, 2, 6, 9, 2, 8, 7, 2, 5, 3238467, 3238467, 3238467, 3238467, 3238467, 3238467, 3238467, 3238467, 3238467, 3238467, 41.699425275238125900, -98.890808159281372100, 72.682359664626079700, -139.252585000287971500, -91.021935444262947200, 285.6730200012627585300, -91.675355856278753500, 56.1336768404475153900, 5.072034371656803800, -40.4198316931202390200
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                         rounds
#> 1                                                1231382, 1231383, 1231384, 1231385, 1231386, 1231387, 1231388, 1231389, 1231390, 1231391, 1231392, 1231393, 1231394, 1231395, 1231396, 1231397, 1231398, 1231399, 1231400, 1231401, 79016, 79016, 79016, 79016, 79016, 79016, 79016, 79016, 79016, 79016, 79016, 79016, 79016, 79016, 79016, 79016, 79016, 79016, 79016, 79016, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19, 20, kills, kills, kills, time, defuse, kills, defuse, defuse, kills, defuse, bomb, kills, kills, kills, kills, kills, kills, kills, kills, kills, 2, 2, 1, 2, 2, 2, 2, 2, 2, 2, 1, 1, 1, 1, 1, 2, 2, 2, 1, 2, default, flawless, default, default, default, default, clutch, default, default, default, default, default, default, default, default, thrifty, default, default, closer, closer, NA, 1, 4, 4, 4, 3, 4, 3, 4, 3, 3, 3, NA, 3, 3, 4, 4, 3, 3, 3, NA, 3, 3, 3, 4, 4, 4, 4, 4, 4, 4, 4, NA, 1, 3, 2, 4, 4, 4, 4
#> 2 1231402, 1231403, 1231404, 1231405, 1231406, 1231407, 1231408, 1231409, 1231410, 1231411, 1231412, 1231413, 1231414, 1231415, 1231416, 1231417, 1231418, 1231419, 1231420, 1231421, 1231422, 79017, 79017, 79017, 79017, 79017, 79017, 79017, 79017, 79017, 79017, 79017, 79017, 79017, 79017, 79017, 79017, 79017, 79017, 79017, 79017, 79017, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19, 20, 21, kills, kills, kills, bomb, kills, kills, defuse, defuse, defuse, kills, kills, bomb, kills, kills, defuse, kills, kills, bomb, kills, defuse, defuse, 1, 1, 1, 1, 1, 2, 2, 2, 2, 2, 1, 1, 2, 1, 1, 2, 1, 2, 1, 1, 1, team_ace, default, default, default, default, thrifty, default, default, default, default, default, default, closer, thrifty, default, default, clutch, thrifty, thrifty, default, default, NA, 3, 3, 4, 4, 4, 4, 4, 4, 4, 3, 4, NA, 1, 4, 4, 4, 4, 2, 4, 4, NA, 1, 3, 2, 4, 3, 4, 4, 4, 4, 4, 4, NA, 3, 2, 4, 4, 2, 4, 2, 3
#> 3      1231423, 1231424, 1231425, 1231426, 1231427, 1231428, 1231429, 1231430, 1231431, 1231432, 1231433, 1231434, 1231435, 1231436, 1231437, 1231438, 1231439, 1231440, 1231441, 1231442, 1231443, 79018, 79018, 79018, 79018, 79018, 79018, 79018, 79018, 79018, 79018, 79018, 79018, 79018, 79018, 79018, 79018, 79018, 79018, 79018, 79018, 79018, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19, 20, 21, kills, kills, bomb, kills, kills, kills, kills, kills, kills, kills, kills, kills, kills, kills, bomb, kills, kills, bomb, kills, kills, kills, 1, 1, 1, 1, 1, 1, 1, 1, 1, 2, 2, 1, 1, 2, 2, 2, 2, 2, 2, 1, 1, default, default, default, flawless, closer, default, default, default, default, thrifty, default, default, default, default, default, default, default, default, default, thrifty, default, NA, 3, 3, 4, 4, 4, 4, 4, 4, 4, 4, 4, NA, 3, 1, 4, 3, 4, 4, 2, 4, NA, 1, 3, 2, 4, 2, 4, 2, 4, 2, 4, 4, NA, 2, 4, 4, 4, 4, 4, 4, 4
# }