Get series

Get series ids for a given event

Usage

get_series(event_id, ..., completed = TRUE, n_results = 50)

Arguments

event_id

Event ID

...

Currently un-used

completed

Whether or not to return completed series

n_results

Number of results to return. 50 is default of the API.

Value

a data.frame

Examples

# \donttest{
get_series(1866) ## VALORANT Champions 2022 - Playoffs
#> Total of 10 possible results.
#>       id eventId team1Id team2Id                startDate bestOf stage bracket
#> 1  35332    1866     603      17 2022-09-13T23:18:44.022Z      3    NA    <NA>
#> 2  35309    1866     286    1625 2022-09-13T14:00:00.000Z      3     2  losers
#> 3  35252    1866    6426     603 2022-09-12T17:00:00.000Z      3     2 winners
#> 4  35246    1866     286    6344 2022-09-12T14:00:00.000Z      3     2 winners
#> 5  35253    1866     224    1625 2022-09-11T17:00:00.000Z      3     1  losers
#> 6  35247    1866      17     738 2022-09-11T14:00:00.000Z      3     1  losers
#> 7  35226    1866     603    1625 2022-09-10T17:00:00.000Z      3     1 winners
#> 8  35225    1866    6426     224 2022-09-10T14:00:00.000Z      3     1 winners
#> 9  35223    1866     286      17 2022-09-09T17:00:00.000Z      3     1 winners
#> 10 35224    1866     738    6344 2022-09-09T14:00:00.000Z      3     1 winners
#>    completed  live winCondition  vlrId
#> 1       TRUE FALSE           NA   <NA>
#> 2       TRUE FALSE           NA 130689
#> 3       TRUE FALSE           NA 130683
#> 4       TRUE FALSE           NA 130682
#> 5       TRUE FALSE           NA 130687
#> 6       TRUE FALSE           NA 130686
#> 7       TRUE FALSE           NA 130681
#> 8       TRUE FALSE           NA 130680
#> 9       TRUE FALSE           NA 130678
#> 10      TRUE FALSE           NA 130679
#>                                                 vodUrl ggbetId pmtStatus
#> 1                                                 <NA>      NA        NA
#> 2  https://www.twitch.tv/videos/1589803546?t=01h08m41s      NA        NA
#> 3  https://www.twitch.tv/videos/1588942223?t=03h38m38s      NA        NA
#> 4  https://www.twitch.tv/videos/1588942223?t=00h58m44s      NA        NA
#> 5  https://www.twitch.tv/videos/1588075240?t=03h46m05s      NA        NA
#> 6  https://www.twitch.tv/videos/1588075240?t=00h29m39s      NA        NA
#> 7                                                 <NA>      NA        NA
#> 8  https://www.twitch.tv/videos/1587106175?t=01h06m30s      NA        NA
#> 9  https://www.twitch.tv/videos/1586150256?t=04h08m03s      NA        NA
#> 10 https://www.twitch.tv/videos/1586150256?t=01h14m22s      NA        NA
#>    pmtRedditUrl pmtJson.valorant
#> 1            NA           posted
#> 2            NA           posted
#> 3            NA           posted
#> 4            NA           posted
#> 5            NA           posted
#> 6            NA           posted
#> 7            NA           posted
#> 8            NA           posted
#> 9            NA           posted
#> 10           NA           posted
#>                                                                                     pmtJson.valorantUrl
#> 1    https://www.reddit.com/r/VALORANT/comments/xdm4ls/xset_vs_funplus_phoenix_valorant_champions_2022/
#> 2     https://www.reddit.com/r/VALORANT/comments/xddi0w/drx_vs_fnatic_valorant_champions_2022_playoffs/
#> 3       https://www.reddit.com/r/VALORANT/comments/xcmtps/optic_gaming_vs_xset_valorant_champions_2022/
#> 4       https://www.reddit.com/r/VALORANT/comments/xcglbm/drx_vs_loud_valorant_champions_2022_playoffs/
#> 5      https://www.reddit.com/r/VALORANT/comments/xbq5kz/team_liquid_vs_fnatic_valorant_champions_2022/
#> 6     https://www.reddit.com/r/VALORANT/comments/xbmh88/funplus_phoenix_vs_leviatán_valorant_champions/
#> 7    https://www.reddit.com/r/VALORANT/comments/xay9j1/xset_vs_fnatic_valorant_champions_2022_playoffs/
#> 8     https://www.reddit.com/r/VALORANT/comments/xaukqe/optic_gaming_vs_team_liquid_valorant_champions/
#> 9     https://www.reddit.com/r/VALORANT/comments/xa42qp/drx_vs_funplus_phoenix_valorant_champions_2022/
#> 10 https://www.reddit.com/r/VALORANT/comments/x9zr0m/leviatán_vs_loud_valorant_champions_2022_playoffs/
#>    pmtJson.valorantcompetitive
#> 1                       posted
#> 2                       posted
#> 3                       posted
#> 4                       posted
#> 5                       posted
#> 6                       posted
#> 7                       posted
#> 8                       posted
#> 9                       posted
#> 10                      posted
#>                                                                                     pmtJson.valorantcompetitiveUrl
#> 1    https://www.reddit.com/r/ValorantCompetitive/comments/xdm4jm/xset_vs_funplus_phoenix_valorant_champions_2022/
#> 2     https://www.reddit.com/r/ValorantCompetitive/comments/xddi1r/drx_vs_fnatic_valorant_champions_2022_playoffs/
#> 3       https://www.reddit.com/r/ValorantCompetitive/comments/xcmtqp/optic_gaming_vs_xset_valorant_champions_2022/
#> 4       https://www.reddit.com/r/ValorantCompetitive/comments/xcglcn/drx_vs_loud_valorant_champions_2022_playoffs/
#> 5      https://www.reddit.com/r/ValorantCompetitive/comments/xbq5lx/team_liquid_vs_fnatic_valorant_champions_2022/
#> 6     https://www.reddit.com/r/ValorantCompetitive/comments/xbmh96/funplus_phoenix_vs_leviatán_valorant_champions/
#> 7    https://www.reddit.com/r/ValorantCompetitive/comments/xay9jr/xset_vs_fnatic_valorant_champions_2022_playoffs/
#> 8     https://www.reddit.com/r/ValorantCompetitive/comments/xaukrj/optic_gaming_vs_team_liquid_valorant_champions/
#> 9     https://www.reddit.com/r/ValorantCompetitive/comments/xa42sc/drx_vs_funplus_phoenix_valorant_champions_2022/
#> 10 https://www.reddit.com/r/ValorantCompetitive/comments/x9zr1i/leviatán_vs_loud_valorant_champions_2022_playoffs/
#>                             eventName                        eventSlug
#> 1  VALORANT Champions 2022 - Playoffs valorant-champions-2022-playoffs
#> 2  VALORANT Champions 2022 - Playoffs valorant-champions-2022-playoffs
#> 3  VALORANT Champions 2022 - Playoffs valorant-champions-2022-playoffs
#> 4  VALORANT Champions 2022 - Playoffs valorant-champions-2022-playoffs
#> 5  VALORANT Champions 2022 - Playoffs valorant-champions-2022-playoffs
#> 6  VALORANT Champions 2022 - Playoffs valorant-champions-2022-playoffs
#> 7  VALORANT Champions 2022 - Playoffs valorant-champions-2022-playoffs
#> 8  VALORANT Champions 2022 - Playoffs valorant-champions-2022-playoffs
#> 9  VALORANT Champions 2022 - Playoffs valorant-champions-2022-playoffs
#> 10 VALORANT Champions 2022 - Playoffs valorant-champions-2022-playoffs
#>                        eventLogoUrl parentEventId         parentEventName
#> 1  https://i.imgur.com/eJZnat0.webp          1858 VALORANT Champions 2022
#> 2  https://i.imgur.com/eJZnat0.webp          1858 VALORANT Champions 2022
#> 3  https://i.imgur.com/eJZnat0.webp          1858 VALORANT Champions 2022
#> 4  https://i.imgur.com/eJZnat0.webp          1858 VALORANT Champions 2022
#> 5  https://i.imgur.com/eJZnat0.webp          1858 VALORANT Champions 2022
#> 6  https://i.imgur.com/eJZnat0.webp          1858 VALORANT Champions 2022
#> 7  https://i.imgur.com/eJZnat0.webp          1858 VALORANT Champions 2022
#> 8  https://i.imgur.com/eJZnat0.webp          1858 VALORANT Champions 2022
#> 9  https://i.imgur.com/eJZnat0.webp          1858 VALORANT Champions 2022
#> 10 https://i.imgur.com/eJZnat0.webp          1858 VALORANT Champions 2022
#>            parentEventSlug team1.id      team1.name team1.shortName
#> 1  valorant-champions-2022      603            XSET            XSET
#> 2  valorant-champions-2022      286             DRX             DRX
#> 3  valorant-champions-2022     6426    OpTic Gaming            OPTC
#> 4  valorant-champions-2022      286             DRX             DRX
#> 5  valorant-champions-2022      224     Team Liquid              TL
#> 6  valorant-champions-2022       17 FunPlus Phoenix             FPX
#> 7  valorant-champions-2022      603            XSET            XSET
#> 8  valorant-champions-2022     6426    OpTic Gaming            OPTC
#> 9  valorant-champions-2022      286             DRX             DRX
#> 10 valorant-champions-2022      738        Leviatán             LEV
#>                                                                                                                                                                                                                                                                                              team1.description
#> 1                                                                                                                                                                                               XSET is an American esports organization, founded in July 2020 with teams in PUBG, Rocket League and VALORANT.
#> 2                                                                                                                                                                                                        Vision Strikers is a Korean VALORANT team formed by a roster of former CS:GO players known as MVP PK.
#> 3                                                                                                                                                                                                                                                                                                             
#> 4                                                                                                                                                                                                        Vision Strikers is a Korean VALORANT team formed by a roster of former CS:GO players known as MVP PK.
#> 5  Team Liquid is a professional esports organization founded in the Netherlands in 2000. Originally a Brood War clan, the team switched to StarCraft II during the SC2 Beta in 2010, and became one of the most successful foreign teams. On August 7, 2020, Team Liquid entered VALORANT by signing fish123.
#> 6                                                                                   FunPlus Phoenix (also known as FunPlus X or FPX) is an esports organization owned by Chinese game developer FunPlus. FunPlus Phoenix now features teams in League of Legends, Fortnite, Call of Duty: Mobile and VALORANT.
#> 7                                                                                                                                                                                               XSET is an American esports organization, founded in July 2020 with teams in PUBG, Rocket League and VALORANT.
#> 8                                                                                                                                                                                                                                                                                                             
#> 9                                                                                                                                                                                                        Vision Strikers is a Korean VALORANT team formed by a roster of former CS:GO players known as MVP PK.
#> 10                                                                                                                                                                                                                                                                                                            
#>                                                             team1.websiteUrl
#> 1                                                          https://xset.com/
#> 2                                                            https://drx.gg/
#> 3                                                    https://opticgaming.gg/
#> 4                                                            https://drx.gg/
#> 5                                                https://www.teamliquid.com/
#> 6  https://web.archive.org/web/20210821045004/https://fpxmw.world.tmall.com/
#> 7                                                          https://xset.com/
#> 8                                                    https://opticgaming.gg/
#> 9                                                            https://drx.gg/
#> 10                                                      https://leviatan.gg/
#>                      team1.logoUrl
#> 1  https://i.imgur.com/TH047Cv.png
#> 2  https://i.imgur.com/nSpWLA0.png
#> 3  https://i.imgur.com/2nXAtuC.png
#> 4  https://i.imgur.com/nSpWLA0.png
#> 5  https://i.imgur.com/vAiJPS0.png
#> 6  https://i.imgur.com/P1YxxFc.png
#> 7  https://i.imgur.com/TH047Cv.png
#> 8  https://i.imgur.com/2nXAtuC.png
#> 9  https://i.imgur.com/nSpWLA0.png
#> 10 https://i.imgur.com/2a8subw.png
#>                                team1.liquipediaUrl
#> 1             https://liquipedia.net/valorant/XSET
#> 2              https://liquipedia.net/valorant/DRX
#> 3     https://liquipedia.net/valorant/OpTic_Gaming
#> 4              https://liquipedia.net/valorant/DRX
#> 5      https://liquipedia.net/valorant/Team_Liquid
#> 6  https://liquipedia.net/valorant/FunPlus_Phoenix
#> 7             https://liquipedia.net/valorant/XSET
#> 8     https://liquipedia.net/valorant/OpTic_Gaming
#> 9              https://liquipedia.net/valorant/DRX
#> 10   https://liquipedia.net/valorant/Leviat%C3%A1n
#>                      team1.twitterUrl                       team1.twitchUrl
#> 1            https://twitter.com/XSET      https://www.twitch.tv/xsetgaming
#> 2          https://twitter.com/DRX_VS          https://www.twitch.tv/drx_vs
#> 3           https://twitter.com/OpTic     https://www.twitch.tv/opticgaming
#> 4          https://twitter.com/DRX_VS          https://www.twitch.tv/drx_vs
#> 5  https://twitter.com/LiquidValorant https://www.twitch.tv/team/teamliquid
#> 6     https://twitter.com/FPX_Esports                                  <NA>
#> 7            https://twitter.com/XSET      https://www.twitch.tv/xsetgaming
#> 8           https://twitter.com/OpTic     https://www.twitch.tv/opticgaming
#> 9          https://twitter.com/DRX_VS          https://www.twitch.tv/drx_vs
#> 10     https://twitter.com/leviatangg                                  <NA>
#>                                   team1.vlrUrl
#> 1             https://www.vlr.gg/team/533/xset
#> 2             https://www.vlr.gg/team/8185/drx
#> 3    https://www.vlr.gg/team/8127/optic-gaming
#> 4             https://www.vlr.gg/team/8185/drx
#> 5      https://www.vlr.gg/team/474/team-liquid
#> 6  https://www.vlr.gg/team/628/funplus-phoenix
#> 7             https://www.vlr.gg/team/533/xset
#> 8    https://www.vlr.gg/team/8127/optic-gaming
#> 9             https://www.vlr.gg/team/8185/drx
#> 10       https://www.vlr.gg/team/2359/leviat-n
#>                                            team1.youtubeUrl
#> 1                        https://www.youtube.com/xsetgaming
#> 2  https://www.youtube.com/channel/UCJN5mgLqkPCTVC5v-12gKjQ
#> 3                        https://www.youtube.com/opticcraft
#> 4  https://www.youtube.com/channel/UCJN5mgLqkPCTVC5v-12gKjQ
#> 5                        https://www.youtube.com/teamliquid
#> 6                    https://www.youtube.com/c/FPX_Official
#> 7                        https://www.youtube.com/xsetgaming
#> 8                        https://www.youtube.com/opticcraft
#> 9  https://www.youtube.com/channel/UCJN5mgLqkPCTVC5v-12gKjQ
#> 10                                                     <NA>
#>            team1.foundedDate team1.countryId team1.regionId team1.rank
#> 1  2020-07-16T05:54:00+00:00             226              2          6
#> 2  2020-06-01T15:26:00+00:00             113              3          2
#> 3  2022-02-10T00:00:00+00:00             226              2          5
#> 4  2020-06-01T15:26:00+00:00             113              3          2
#> 5  2020-08-07T15:02:00+00:00             252              1          8
#> 6  2017-12-21T06:47:00+00:00             252              1          1
#> 7  2020-07-16T05:54:00+00:00             226              2          6
#> 8  2022-02-10T00:00:00+00:00             226              2          5
#> 9  2020-06-01T15:26:00+00:00             113              3          2
#> 10 2020-06-02T00:00:00+00:00              43              4          9
#>    team1.regionRank            team1.aliases team2.id      team2.name
#> 1                 2                     XSET       17 FunPlus Phoenix
#> 2                 1                      DRX     1625          FNATIC
#> 3                 1                    OpTic      603            XSET
#> 4                 1                      DRX     6344            LOUD
#> 5                 3               TL, LIQUID     1625          FNATIC
#> 6                 1             FPX, FunPlus      738        Leviatán
#> 7                 2                     XSET     1625          FNATIC
#> 8                 1                    OpTic      224     Team Liquid
#> 9                 1                      DRX       17 FunPlus Phoenix
#> 10                2 Leviatán, Leviatan, Levi     6344            LOUD
#>    team2.shortName
#> 1              FPX
#> 2              FNC
#> 3             XSET
#> 4             LOUD
#> 5              FNC
#> 6              LEV
#> 7              FNC
#> 8               TL
#> 9              FPX
#> 10            LOUD
#>                                                                                                                                                                                                                                                                                              team2.description
#> 1                                                                                   FunPlus Phoenix (also known as FunPlus X or FPX) is an esports organization owned by Chinese game developer FunPlus. FunPlus Phoenix now features teams in League of Legends, Fortnite, Call of Duty: Mobile and VALORANT.
#> 2                                                                                                                                                                                                                                                                                                             
#> 3                                                                                                                                                                                               XSET is an American esports organization, founded in July 2020 with teams in PUBG, Rocket League and VALORANT.
#> 4                                                                                                                                                                                                                                                                                                             
#> 5                                                                                                                                                                                                                                                                                                             
#> 6                                                                                                                                                                                                                                                                                                             
#> 7                                                                                                                                                                                                                                                                                                             
#> 8  Team Liquid is a professional esports organization founded in the Netherlands in 2000. Originally a Brood War clan, the team switched to StarCraft II during the SC2 Beta in 2010, and became one of the most successful foreign teams. On August 7, 2020, Team Liquid entered VALORANT by signing fish123.
#> 9                                                                                   FunPlus Phoenix (also known as FunPlus X or FPX) is an esports organization owned by Chinese game developer FunPlus. FunPlus Phoenix now features teams in League of Legends, Fortnite, Call of Duty: Mobile and VALORANT.
#> 10                                                                                                                                                                                                                                                                                                            
#>                                                             team2.websiteUrl
#> 1  https://web.archive.org/web/20210821045004/https://fpxmw.world.tmall.com/
#> 2                                                        https://fnatic.com/
#> 3                                                          https://xset.com/
#> 4                                                           https://loud.gg/
#> 5                                                        https://fnatic.com/
#> 6                                                       https://leviatan.gg/
#> 7                                                        https://fnatic.com/
#> 8                                                https://www.teamliquid.com/
#> 9  https://web.archive.org/web/20210821045004/https://fpxmw.world.tmall.com/
#> 10                                                          https://loud.gg/
#>                      team2.logoUrl
#> 1  https://i.imgur.com/P1YxxFc.png
#> 2  https://i.imgur.com/CGAJxmz.png
#> 3  https://i.imgur.com/TH047Cv.png
#> 4  https://i.imgur.com/AiZFd1I.png
#> 5  https://i.imgur.com/CGAJxmz.png
#> 6  https://i.imgur.com/2a8subw.png
#> 7  https://i.imgur.com/CGAJxmz.png
#> 8  https://i.imgur.com/vAiJPS0.png
#> 9  https://i.imgur.com/P1YxxFc.png
#> 10 https://i.imgur.com/AiZFd1I.png
#>                                team2.liquipediaUrl
#> 1  https://liquipedia.net/valorant/FunPlus_Phoenix
#> 2           https://liquipedia.net/valorant/Fnatic
#> 3             https://liquipedia.net/valorant/XSET
#> 4             https://liquipedia.net/valorant/LOUD
#> 5           https://liquipedia.net/valorant/Fnatic
#> 6    https://liquipedia.net/valorant/Leviat%C3%A1n
#> 7           https://liquipedia.net/valorant/Fnatic
#> 8      https://liquipedia.net/valorant/Team_Liquid
#> 9  https://liquipedia.net/valorant/FunPlus_Phoenix
#> 10            https://liquipedia.net/valorant/LOUD
#>                      team2.twitterUrl                       team2.twitchUrl
#> 1     https://twitter.com/FPX_Esports                                  <NA>
#> 2          https://twitter.com/FNATIC          https://www.twitch.tv/fnatic
#> 3            https://twitter.com/XSET      https://www.twitch.tv/xsetgaming
#> 4          https://twitter.com/loudgg            https://www.twitch.tv/loud
#> 5          https://twitter.com/FNATIC          https://www.twitch.tv/fnatic
#> 6      https://twitter.com/leviatangg                                  <NA>
#> 7          https://twitter.com/FNATIC          https://www.twitch.tv/fnatic
#> 8  https://twitter.com/LiquidValorant https://www.twitch.tv/team/teamliquid
#> 9     https://twitter.com/FPX_Esports                                  <NA>
#> 10         https://twitter.com/loudgg            https://www.twitch.tv/loud
#>                                   team2.vlrUrl
#> 1  https://www.vlr.gg/team/628/funplus-phoenix
#> 2          https://www.vlr.gg/team/2593/fnatic
#> 3             https://www.vlr.gg/team/533/xset
#> 4            https://www.vlr.gg/team/6961/loud
#> 5          https://www.vlr.gg/team/2593/fnatic
#> 6        https://www.vlr.gg/team/2359/leviat-n
#> 7          https://www.vlr.gg/team/2593/fnatic
#> 8      https://www.vlr.gg/team/474/team-liquid
#> 9  https://www.vlr.gg/team/628/funplus-phoenix
#> 10           https://www.vlr.gg/team/6961/loud
#>                                            team2.youtubeUrl
#> 1                    https://www.youtube.com/c/FPX_Official
#> 2                          https://www.youtube.com/c/fnatic
#> 3                        https://www.youtube.com/xsetgaming
#> 4  https://www.youtube.com/channel/UC7iwNp4GUynlGXvK-6KD0Rw
#> 5                          https://www.youtube.com/c/fnatic
#> 6                                                      <NA>
#> 7                          https://www.youtube.com/c/fnatic
#> 8                        https://www.youtube.com/teamliquid
#> 9                    https://www.youtube.com/c/FPX_Official
#> 10 https://www.youtube.com/channel/UC7iwNp4GUynlGXvK-6KD0Rw
#>            team2.foundedDate team2.countryId team2.regionId team2.rank
#> 1  2017-12-21T06:47:00+00:00             252              1          1
#> 2  2021-02-03T00:00:00+00:00             252              1          7
#> 3  2020-07-16T05:54:00+00:00             226              2          6
#> 4  2022-02-01T00:00:00+00:00              30              4          4
#> 5  2021-02-03T00:00:00+00:00             252              1          7
#> 6  2020-06-02T00:00:00+00:00              43              4          9
#> 7  2021-02-03T00:00:00+00:00             252              1          7
#> 8  2020-08-07T15:02:00+00:00             252              1          8
#> 9  2017-12-21T06:47:00+00:00             252              1          1
#> 10 2022-02-01T00:00:00+00:00              30              4          4
#>    team2.regionRank            team2.aliases
#> 1                 1             FPX, FunPlus
#> 2                 2                 FNC, FNA
#> 3                 2                     XSET
#> 4                 1                     NULL
#> 5                 2                 FNC, FNA
#> 6                 2 Leviatán, Leviatan, Levi
#> 7                 2                 FNC, FNA
#> 8                 3               TL, LIQUID
#> 9                 1             FPX, FunPlus
#> 10                1                     NULL
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                               matches
#> 1                                                                                                                                                             79261, 79262, 79263, 26, 26, 26, 35332, 35332, 35332, 1, 2, 3, 2e4783dd-39a8-48f3-8957-94bcb38abdd6, 96a6b1cc-5977-4b85-89e1-c62b6212e22b, a0bd63ad-34d0-42b0-8409-c430d56d983b, 10, 3, 1, 2022-09-13T17:55:05.728Z, 2022-09-13T19:18:16.354Z, 2022-09-13T22:44:01.213Z, 4417422, 3525494, 2626680, TRUE, TRUE, TRUE, 3e47230a-463c-a301-eb7d-67bb60357d4f, 3e47230a-463c-a301-eb7d-67bb60357d4f, 3e47230a-463c-a301-eb7d-67bb60357d4f, /Game/GameModes/Bomb/BombGameMode.BombGameMode_C, /Game/GameModes/Bomb/BombGameMode.BombGameMode_C, /Game/GameModes/Bomb/BombGameMode.BombGameMode_C, esports, esports, esports, FALSE, FALSE, FALSE, TRUE, TRUE, TRUE, 1, 2, 2, 1, 2, 2, 1, 2, 2, standard, standard, standard, FALSE, FALSE, FALSE, 13, 10, 14, 11, 13, 16, NA, NA, NA, NA, NA, NA, 10, 3, 1, Pearl, Bind, Ascent, /Game/Maps/Pitt/Pitt, /Game/Maps/Duality/Duality, /Game/Maps/Ascent/Ascent, 3062, 494, 1802, 3582, 3346, 3582, 3062, 3346, 494, 1802, 3582, 3062, 3346, 1802, 494, 2225, 1256, 2359, 2335, 3043, 1256, 2225, 2335, 3043, 2359, 2225, 2335, 3043, 1256, 2359, 15, 16, 17, 18, 19, 2, 6, 13, 15, 17, 4, 5, 11, 12, 16, 9, 13, 15, 16, 17, 2, 6, 8, 16, 19, 9, 11, 12, 16, 19
#> 2                   79190, 79191, 79192, 26, 26, 26, 35309, 35309, 35309, 1, 2, 3, 2ba41de5-2a99-471e-bede-ec09a55fa0d9, 1bb55ce7-b57e-497e-a071-6d0cd134bc3e, 509274fb-31a6-4f90-8fbc-1d4ebb056b22, 4, 1, 9, 2022-09-13T13:56:53.891Z, 2022-09-13T15:08:51.741Z, 2022-09-13T16:29:14.810Z, 3805759, 4204457, 3111702, TRUE, TRUE, TRUE, 3e47230a-463c-a301-eb7d-67bb60357d4f, 3e47230a-463c-a301-eb7d-67bb60357d4f, 3e47230a-463c-a301-eb7d-67bb60357d4f, /Game/GameModes/Bomb/BombGameMode.BombGameMode_C, /Game/GameModes/Bomb/BombGameMode.BombGameMode_C, /Game/GameModes/Bomb/BombGameMode.BombGameMode_C, esports, esports, esports, FALSE, FALSE, FALSE, TRUE, TRUE, TRUE, 2, 1, 2, 2, 1, 2, 2, 1, 1, standard, standard, standard, FALSE, FALSE, FALSE, 7, 14, 13, 13, 12, 9, 95145, 95146, 95147, https://www.twitch.tv/videos/1589803546?t=01h13m37s, https://www.twitch.tv/videos/1589803546?t=02h14m00s, https://www.twitch.tv/videos/1589803546?t=03h33m06s, 4, 1, 9, Icebox, Ascent, Fracture, /Game/Maps/Port/Port, /Game/Maps/Ascent/Ascent, /Game/Maps/Canyon/Canyon, 814, 1550, 452, 582, 1251, 814, 1251, 452, 1550, 582, 582, 452, 1251, 1550, 814, 718, 14543, 2011, 76, 2937, 14543, 2011, 2937, 76, 718, 718, 2011, 76, 2937, 14543, 4, 5, 6, 9, 12, 4, 12, 15, 16, 19, 1, 8, 17, 18, 19, 4, 5, 6, 9, 17, 5, 11, 12, 16, 19, 1, 8, 16, 17, 18
#> 3        79079, 79080, 79081, 26, 26, 26, 35252, 35252, 35252, 1, 2, 3, fff670c7-e775-4105-b3bf-82c2e1e6d4c7, ac2a608f-e8a2-4e94-afaf-ee567ab26d14, 3be5d0c2-0083-4c63-bbab-8ad0cae3f8a7, 7, 1, 10, 2022-09-12T16:29:27.308Z, 2022-09-12T17:45:03.310Z, 2022-09-12T18:51:41.636Z, 4018488, 3364515, 4707604, TRUE, TRUE, TRUE, 3e47230a-463c-a301-eb7d-67bb60357d4f, 3e47230a-463c-a301-eb7d-67bb60357d4f, 3e47230a-463c-a301-eb7d-67bb60357d4f, /Game/GameModes/Bomb/BombGameMode.BombGameMode_C, /Game/GameModes/Bomb/BombGameMode.BombGameMode_C, /Game/GameModes/Bomb/BombGameMode.BombGameMode_C, esports, esports, esports, FALSE, FALSE, FALSE, TRUE, TRUE, TRUE, 2, 1, 1, 2, 1, 1, 1, 2, 1, standard, standard, standard, FALSE, FALSE, FALSE, 13, 11, 13, 9, 13, 8, 95125, 95126, 95127, https://www.twitch.tv/videos/1588942223?t=03h45m13s, https://www.twitch.tv/videos/1588942223?t=04h49m56s, https://www.twitch.tv/videos/1588942223?t=05h59m23s, 7, 1, 10, Haven, Ascent, Pearl, /Game/Maps/Triad/Triad, /Game/Maps/Ascent/Ascent, /Game/Maps/Pitt/Pitt, 2841, 3211, 398, 2716, 2483, 2483, 398, 3211, 2716, 2841, 2841, 398, 3211, 2716, 2483, 494, 3582, 3062, 1802, 3346, 3582, 3062, 3346, 1802, 494, 3062, 494, 1802, 3582, 3346, 5, 7, 11, 17, 19, 4, 15, 16, 17, 19, 6, 15, 16, 17, 19, 1, 2, 11, 17, 19, 4, 5, 11, 12, 16, 15, 16, 17, 18, 19
#> 4                                                                                                                                                                                                                                                                                                               79061, 79062, 79063, 26, 26, NA, 35246, 35246, 35246, 1, 2, 3, ed03901e-ba05-44fa-84cc-fa7f2c3062ad, 20ae39c9-1cc1-4d31-be3a-74d7a79d4c4e, NA, 8, 7, 4, 2022-09-12T13:57:18.981Z, 2022-09-12T15:19:39.685Z, NA, 4292331, 2376021, NA, TRUE, TRUE, FALSE, 3e47230a-463c-a301-eb7d-67bb60357d4f, 3e47230a-463c-a301-eb7d-67bb60357d4f, NA, /Game/GameModes/Bomb/BombGameMode.BombGameMode_C, /Game/GameModes/Bomb/BombGameMode.BombGameMode_C, NA, esports, esports, NA, FALSE, FALSE, FALSE, TRUE, TRUE, FALSE, 1, 1, NA, 1, 1, NA, 2, 2, NA, standard, standard, NA, FALSE, FALSE, FALSE, 11, 6, NA, 13, 13, NA, 95122, 95123, 95124, https://www.twitch.tv/videos/1588942223?t=01h16m02s, https://www.twitch.tv/videos/1588942223?t=02h21m33s, NA, 8, 7, 4, Breeze, Haven, Icebox, /Game/Maps/Foxtrot/Foxtrot, /Game/Maps/Triad/Triad, /Game/Maps/Port/Port, 814, 452, 1251, 582, 1550, 582, 1550, 452, 1251, 814, 4746, 2294, 6037, 4748, 2895, 2895, 4748, 6037, 4746, 2294, 4, 6, 12, 13, 17, 1, 5, 11, 12, 19, 3, 4, 6, 12, 16, 1, 2, 11, 17, 19
#> 5                                                                                                                                                                                                                                                                                                                                    79082, 79083, 79084, 26, 26, NA, 35253, 35253, 35253, 1, 2, 3, d7992b50-b185-426b-ad3f-1d52940192fc, 3e404f3a-b1ff-4f96-965d-dbb8014039f8, NA, 3, 7, NA, 2022-09-11T16:41:13.036Z, 2022-09-11T17:51:47.279Z, NA, 3620886, 2823323, NA, TRUE, TRUE, FALSE, 3e47230a-463c-a301-eb7d-67bb60357d4f, 3e47230a-463c-a301-eb7d-67bb60357d4f, NA, /Game/GameModes/Bomb/BombGameMode.BombGameMode_C, /Game/GameModes/Bomb/BombGameMode.BombGameMode_C, NA, esports, esports, NA, FALSE, FALSE, FALSE, TRUE, TRUE, FALSE, 2, 2, NA, 2, 2, NA, 2, 2, NA, standard, standard, NA, FALSE, FALSE, FALSE, 6, 7, NA, 13, 13, NA, 95139, 95140, 95141, https://www.twitch.tv/videos/1588075240?t=03h57m19s, https://www.twitch.tv/videos/1588075240?t=04h56m21s, NA, 3, 7, NA, Bind, Haven, NA, /Game/Maps/Duality/Duality, /Game/Maps/Triad/Triad, NA, 1690, 17788, 750, 10211, 2056, 17788, 1690, 750, 10211, 2056, 14543, 718, 2011, 76, 2937, 76, 14543, 2011, 2937, 718, 2, 6, 8, 9, 13, 1, 2, 11, 17, 19, 2, 6, 8, 13, 17, 1, 2, 11, 17, 19
#> 6                                                                                                                                                                                                                                                                                                     79064, 79065, 79066, 26, 26, NA, 35247, 35247, 35247, 1, 2, 3, 6a702bf4-2153-4306-acbd-aa41b9309663, 619e9f83-06a5-442d-81c4-3716bfa4f756, NA, 4, 1, 3, 2022-09-11T13:56:40.742Z, 2022-09-11T15:09:37.530Z, NA, 3767748, 3621468, NA, TRUE, TRUE, FALSE, 3e47230a-463c-a301-eb7d-67bb60357d4f, 3e47230a-463c-a301-eb7d-67bb60357d4f, NA, /Game/GameModes/Bomb/BombGameMode.BombGameMode_C, /Game/GameModes/Bomb/BombGameMode.BombGameMode_C, NA, esports, esports, NA, FALSE, FALSE, FALSE, TRUE, TRUE, FALSE, 1, 1, NA, 1, 1, NA, 1, 1, NA, standard, standard, NA, FALSE, FALSE, FALSE, 13, 13, NA, 6, 11, NA, 95136, 95137, 95138, https://www.twitch.tv/videos/1588075240?t=01h17m25s, https://www.twitch.tv/videos/1588075240?t=02h14m00s, NA, 4, 1, 3, Icebox, Ascent, Bind, /Game/Maps/Port/Port, /Game/Maps/Ascent/Ascent, /Game/Maps/Duality/Duality, 1256, 2225, 2359, 3043, 2335, 2225, 2335, 3043, 1256, 2359, 3531, 6414, 3421, 21588, 3532, 3531, 3421, 6414, 3532, 21588, 2, 6, 9, 17, 19, 3, 11, 12, 16, 19, 4, 6, 9, 16, 17, 4, 15, 16, 17, 19
#> 7                                                                                                                                                                                                                                                                                                                                                                                                                                79019, 79020, 79021, 26, 26, NA, 35226, 35226, 35226, 1, 2, 3, 721d4efa-e802-4b9f-bc3f-8238f40b62e8, 92901b1a-79d5-4295-a134-e59dfdf75bf4, NA, 10, 9, NA, 2022-09-10T17:38:59.080Z, 2022-09-10T18:51:52.782Z, NA, 3799095, 3236742, NA, TRUE, TRUE, FALSE, 3e47230a-463c-a301-eb7d-67bb60357d4f, 3e47230a-463c-a301-eb7d-67bb60357d4f, NA, /Game/GameModes/Bomb/BombGameMode.BombGameMode_C, /Game/GameModes/Bomb/BombGameMode.BombGameMode_C, NA, esports, esports, NA, FALSE, FALSE, FALSE, TRUE, TRUE, FALSE, 2, 2, NA, 2, 2, NA, 1, 1, NA, standard, standard, NA, FALSE, FALSE, FALSE, 13, 13, NA, 8, 11, NA, 95119, 95120, 95121, NA, NA, NA, 10, 9, NA, Pearl, Fracture, NA, /Game/Maps/Pitt/Pitt, /Game/Maps/Canyon/Canyon, NA, 3062, 494, 1802, 3582, 3346, 494, 3062, 3346, 1802, 3582, 76, 2011, 2937, 14543, 718, 718, 2011, 76, 2937, 14543, 15, 16, 17, 18, 19, 1, 9, 11, 17, 18, 13, 15, 17, 18, 19, 1, 8, 16, 17, 18
#> 8  79016, 79017, 79018, 26, 26, 26, 35225, 35225, 35225, 1, 2, 3, a90c6bfb-e656-4d0f-ab32-fa1c275374b4, f2c5ed14-6718-4219-9898-2fd8802f587f, 15ef3a70-7e55-4034-9cae-f6a349942daf, 8, 3, 1, 2022-09-10T13:57:37.704Z, 2022-09-10T15:08:28.124Z, 2022-09-10T16:15:29.650Z, 3708997, 3338271, 3238467, TRUE, TRUE, TRUE, 3e47230a-463c-a301-eb7d-67bb60357d4f, 3e47230a-463c-a301-eb7d-67bb60357d4f, 3e47230a-463c-a301-eb7d-67bb60357d4f, /Game/GameModes/Bomb/BombGameMode.BombGameMode_C, /Game/GameModes/Bomb/BombGameMode.BombGameMode_C, /Game/GameModes/Bomb/BombGameMode.BombGameMode_C, esports, esports, esports, FALSE, FALSE, FALSE, TRUE, TRUE, TRUE, 1, 1, 1, 1, 1, 1, 2, 1, 1, standard, standard, standard, FALSE, FALSE, FALSE, 7, 13, 13, 13, 8, 8, 95116, 95117, 95118, https://www.twitch.tv/videos/1587106175?t=01h16m05s, https://www.twitch.tv/videos/1587106175?t=02h13m04s, https://www.twitch.tv/videos/1587106175?t=03h18m54s, 8, 3, 1, Breeze, Bind, Ascent, /Game/Maps/Foxtrot/Foxtrot, /Game/Maps/Duality/Duality, /Game/Maps/Ascent/Ascent, 398, 2483, 2841, 2716, 3211, 3211, 2841, 398, 2483, 2716, 2483, 398, 3211, 2716, 2841, 750, 17788, 1690, 2056, 10211, 1690, 17788, 750, 10211, 2056, 1690, 750, 17788, 10211, 2056, 1, 4, 6, 17, 18, 2, 6, 8, 13, 17, 4, 15, 16, 17, 19, 4, 6, 7, 16, 17, 2, 6, 8, 9, 13, 7, 11, 16, 17, 19
#> 9                                                                                                                                                                                                                                                                                                                             79010, 79011, 79012, 26, 26, NA, 35223, 35223, 35223, 1, 2, 3, b8838375-39f1-4b3d-a1fa-02dc2e184df0, 080c51d4-9bdf-4e63-be96-508c66b9a492, NA, 1, 8, NA, 2022-09-09T16:49:25.356Z, 2022-09-09T18:35:23.523Z, NA, 5813959, 2473677, NA, TRUE, TRUE, FALSE, 3e47230a-463c-a301-eb7d-67bb60357d4f, 3e47230a-463c-a301-eb7d-67bb60357d4f, NA, /Game/GameModes/Bomb/BombGameMode.BombGameMode_C, /Game/GameModes/Bomb/BombGameMode.BombGameMode_C, NA, esports, esports, NA, FALSE, FALSE, FALSE, TRUE, TRUE, FALSE, 2, 1, NA, 2, 1, NA, 1, 1, NA, standard, standard, NA, FALSE, FALSE, FALSE, 19, 13, NA, 17, 5, NA, 95110, 95111, 95112, https://www.twitch.tv/videos/1586150256?t=04h08m03s, https://www.twitch.tv/videos/1586150256?t=05h41m02s, NA, 1, 8, NA, Ascent, Breeze, NA, /Game/Maps/Ascent/Ascent, /Game/Maps/Foxtrot/Foxtrot, NA, 814, 1251, 452, 1550, 582, 814, 452, 1251, 582, 1550, 2225, 2335, 1256, 3043, 2359, 2359, 2225, 2335, 3043, 1256, 4, 12, 15, 16, 19, 4, 6, 12, 13, 17, 9, 11, 16, 17, 19, 4, 6, 11, 12, 16
#> 10                                                                                                                                                                                                                                                                                                                                 79013, 79014, 79015, 26, 26, NA, 35224, 35224, 35224, 1, 2, 3, d4db1c16-0bf7-4bdc-9b08-3725e73c6a7b, fc86b057-9a36-44e3-ae7b-f073e8185975, NA, 4, 7, NA, 2022-09-09T13:57:09.014Z, 2022-09-09T15:33:54.141Z, NA, 5225539, 2726821, NA, TRUE, TRUE, FALSE, 3e47230a-463c-a301-eb7d-67bb60357d4f, 3e47230a-463c-a301-eb7d-67bb60357d4f, NA, /Game/GameModes/Bomb/BombGameMode.BombGameMode_C, /Game/GameModes/Bomb/BombGameMode.BombGameMode_C, NA, esports, esports, NA, FALSE, FALSE, FALSE, TRUE, TRUE, FALSE, 2, 1, NA, 2, 1, NA, 2, 2, NA, standard, standard, NA, FALSE, FALSE, FALSE, 12, 6, NA, 14, 13, NA, 95113, 95114, 95115, https://www.twitch.tv/videos/1586150256?t=01h17m00s, https://www.twitch.tv/videos/1586150256?t=02h39m07s, NA, 4, 7, NA, Icebox, Haven, NA, /Game/Maps/Port/Port, /Game/Maps/Triad/Triad, NA, 3531, 6414, 3421, 21588, 3532, 3531, 6414, 3421, 3532, 21588, 2294, 4746, 6037, 2895, 4748, 2895, 4748, 6037, 4746, 2294, 4, 6, 9, 16, 17, 1, 2, 11, 17, 19, 4, 6, 9, 16, 17, 1, 2, 11, 17, 19
# }