Get events

Get event ids given a free-text query

Usage

get_events(
  query = NULL,
  ...,
  sort_by = "startDate",
  ascending = FALSE,
  has_series = TRUE,
  n_results = 50
)

Arguments

query

Text description of event, e.g. "VALORANT Champions 2022 Playoffs"

...

Currently un-used

sort_by, ascending

How to return the results

has_series

Whether or not to return events with completed series

n_results

Number of results to return. API tries to respect n_results even if event name doesn't match query. 50 is default of the API.

Value

a data.frame

Details

Note that sometimes the query doesn't seem to work. Results will be returned, but they are a default set of results.

Examples

# \donttest{
events <- get_events("VALORANT Champions 2022")
dplyr::glimpse(events)
#> Rows: 19
#> Columns: 29
#> $ id                <int> 1866, 1865, 1858, 1694, 1623, 1687, 1351, 1335, 1346…
#> $ name              <chr> "VALORANT Champions 2022 - Playoffs", "VALORANT Cham…
#> $ shortName         <chr> "Playoffs", "Group Stage", "Champions 2022", "Playof…
#> $ description       <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, …
#> $ logoUrl           <chr> NA, NA, "https://i.imgur.com/eJZnat0.webp", NA, "htt…
#> $ regionId          <int> 7, 7, 7, 7, 7, 7, 7, 7, 7, 3, 3, 7, 7, 7, 7, 7, 7, 7…
#> $ countryId         <int> 218, 218, 218, 58, 58, 252, 98, 98, 98, NA, NA, 252,…
#> $ startDate         <chr> "2022-09-09T15:46:00.000Z", "2022-08-31T15:36:00.000…
#> $ endDate           <chr> "2022-09-19T15:46:00.000Z", "2022-09-09T15:36:00.000…
#> $ prizePool         <int> NA, NA, 1000000, NA, 650000, NA, NA, 675000, NA, NA,…
#> $ prizePoolCurrency <chr> "USD", "USD", "USD", "USD", "USD", "USD", "USD", "US…
#> $ url               <chr> "https://liquipedia.net/valorant/VCT/2022/Champions"…
#> $ imageUrl          <chr> NA, NA, "https://i.imgur.com/eJZnat0.webp", NA, "htt…
#> $ winnerStageCount  <int> NA, NA, NA, NA, NA, NA, NA, NA, NA, 3, NA, NA, NA, N…
#> $ loserStageCount   <int> NA, NA, NA, NA, NA, NA, NA, NA, NA, 2, NA, NA, NA, N…
#> $ live              <lgl> FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FAL…
#> $ rank              <int> 9, 9, 9, 9, 9, 9, 9, 9, 9, 5, 5, 9, 9, 9, 9, 9, 10, …
#> $ bracketJson       <df[,4]> <data.frame[19 x 4]>
#> $ pmtJson           <lgl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, N…
#> $ parent            <lgl> FALSE, FALSE, TRUE, FALSE, TRUE, FALSE, FALSE, TRUE,…
#> $ parentId          <int> 1858, 1858, 1760, 1623, 1763, 1623, 1335, 1761, 1335…
#> $ childLabel        <chr> "Playoffs", "Group Stage", "Champions 2022", "Playof…
#> $ keywords          <list> [], [], [], [], [], [], [], [], [], [], [], [], [], …
#> $ slug              <chr> "valorant-champions-2022-playoffs", "valorant-champ…
#> $ seriesCount       <int> 14, 20, 0, 14, 0, 10, 14, 0, 10, 6, 0, 0, 0, 7, 0, 2…
#> $ importance        <int> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0…
#> $ type              <chr> "bracket", "bracket", "event", "bracket", "event", "…
#> $ region            <df[,4]> <data.frame[19 x 4]>
#> $ country           <df[,6]> <data.frame[19 x 6]>
# }